refactor: Refactored abilities
