+++ Assets/_WorldsCollide/Scripts/Abilities/AB_Thrust.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "Ability_Thrust", menuName = "ScriptableObjects/Abilities/Rogue/Thrust")]
-public class AB_Thrust : PlayerAbility
+public class AB_Thrust : Ability
 {
 
     public AB_Thrust(GameObject caster): base(caster)
